{"types":"C","runs":17}

+++ Problem49.java
@@ -4,12 +4,12 @@ public class Problem49 extends ConsoleProgram {
 	
 	public void run() {
 		String numStr = readLine("Enter a number: ");
-		double finalNum = stringToInteger(numStr);
+		int finalNum = stringToInteger(numStr);
 		println(finalNum);
 	}
 
-	private double stringToInteger(String numStr) {
-		double num = 0;
+	private int stringToInteger(String numStr) {
+		int num = 0;
 		for(int i = 0; i < numStr.length(); i++){
 			int currDigit = charToInteger(numStr.charAt(i));
 			num = num * 10 + currDigit;
